+++ src/main/java/org/java/onion/chapter2/practice/CalculatorV1Test.java
@@ -1,11 +1,11 @@
-package org.java.onion.chapter2.testandrefactor;
+package org.java.onion.chapter2.practice;
 
 /*
 테스트 코드를 담당하는 별도의 클래스를 추가했으나,
 main()메서드 하나에 프로덕션 코드의 여러 메소드를 동시에 테스트함.
 이는 복잡도도 높아지고 유지하기에 부담이 있다.
  */
-public class CalculatorTest {
+public class CalculatorV1Test {
     public static void main(String[] args) {
         CalculatorV2 cal = new CalculatorV2();
         System.out.println(cal.add(3,4));
