+++ src/main/java/org/java/onion/chapter2/practice/CalculatorV3Test.java
@@ -1,4 +1,4 @@
-package org.java.onion.chapter2.testandrefactor;
+package org.java.onion.chapter2.practice;
 
 
 import org.junit.jupiter.api.Test;
